+++ lib/root.ex
@@ -13,8 +13,8 @@ defmodule Tamnoon.Components.Root do
         <meta charset="utf-8">
         <title>Tamnoon App</title>
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <script src="tamnoon_dom.js"></script>
-        <script src="tamnoon_driver.js"></script>
+        <script src="tamnoon/tamnoon_dom.js"></script>
+        <script src="tamnoon/tamnoon_driver.js"></script>
       </head>
 
       <body>