+++ changelog.md
@@ -10,6 +10,8 @@
 
 - Changed methods to accept the `"value"` key for the value instead of `"val"`.
 
+- Fixed a bug where after setting an element's `disabled` attribute to a state value, changes to it wouldn't re-enable the element properly.
+
 - Changed the `root` component that gets generated with `m:Mix.Tasks.Tamnoon.OverrideRoot` to not try rendering an `app.html.heex` component by default.