+++ changelog.md
@@ -16,6 +16,8 @@
 
 - Changed the `root` component that gets generated with `m:Mix.Tasks.Tamnoon.OverrideRoot` to not try rendering an `app.html.heex` component by default.
 
+- Fixed a bug causing `<a>` elements to not send their value.
+
 
 ### v1.0.0-a.5 (13.06.25)